requirements.txt
asgiref==3.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
defusedxml==0.7.1
Django==3.1.8
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==21.2.0
httpie==3.2.2
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
packaging==23.1
Pillow==10.0.0
Pygments==2.15.1
PyJWT==2.7.0
PySocks==1.7.1
pytz==2021.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.2
sqlparse==0.4.1
urllib3==2.0.3
whitenoise==6.5.0
